requirements.txt
pandas>=2.1.0
streamlit>=1.37.0
python-dotenv>=1.0.0
langchain>=0.2.0
langchain-openai>=0.1.7
pydantic>=2.7.0
